Add checked exceptions for `run`
We can provide similar behavior to `Control.Monad.Except.Checked`
except in the world of `run`.
The general idea is the same though:
use polymorphic variants to encode exceptions and handle appropriately.

This is mostly a copy of `Control.Monad.Except.Checked`.
There's been some massaging of types and values to make things line up.